Changed location of Nbt converter to SpigotReflectionUtil
